Revert "added client class WieldableSystem"
This reverts commit 7200c4d286b69172f543ff43a2a224d0f587785d.
