+++ Content.Shared/Weapons/Melee/Components/MeleeRequiresWieldComponent.cs
@@ -6,7 +6,7 @@ namespace Content.Shared.Weapons.Melee.Components;
 /// <summary>
 /// Indicates that this meleeweapon requires wielding to be useable.
 /// </summary>
-[RegisterComponent, NetworkedComponent, Access(typeof(SharedWieldableSystem))]
+[RegisterComponent, NetworkedComponent, Access(typeof(WieldableSystem))]
 public sealed partial class MeleeRequiresWieldComponent : Component
 {
