+++ Content.Shared/Weapons/Ranged/Components/GunRequiresWieldComponent.cs
@@ -6,7 +6,7 @@ namespace Content.Shared.Weapons.Ranged.Components;
 /// <summary>
 /// Indicates that this gun requires wielding to be useable.
 /// </summary>
-[RegisterComponent, NetworkedComponent, Access(typeof(SharedWieldableSystem))]
+[RegisterComponent, NetworkedComponent, Access(typeof(WieldableSystem))]
 public sealed partial class GunRequiresWieldComponent : Component
 {
