+++ Content.Shared/Wieldable/Components/IncreaseDamageOnWieldComponent.cs
@@ -2,7 +2,7 @@ using Content.Shared.Damage;
 
 namespace Content.Shared.Wieldable.Components;
 
-[RegisterComponent, Access(typeof(SharedWieldableSystem))]
+[RegisterComponent, Access(typeof(WieldableSystem))]
 public sealed partial class IncreaseDamageOnWieldComponent : Component
 {
     [DataField("damage", required: true)]
