+++ Content.Shared/Wieldable/Components/WieldableComponent.cs
@@ -9,7 +9,7 @@ namespace Content.Shared.Wieldable.Components;
 /// <summary>
 ///     Used for objects that can be wielded in two or more hands,
 /// </summary>
-[RegisterComponent, NetworkedComponent, Access(typeof(SharedWieldableSystem)), AutoGenerateComponentState]
+[RegisterComponent, NetworkedComponent, Access(typeof(WieldableSystem)), AutoGenerateComponentState]
 public sealed partial class WieldableComponent : Component
 {
     [DataField("wieldSound")]
