--- Content.Shared/Wieldable/TwoHandWieldingActionEvent.cs
@@ -1,5 +1,0 @@
-using Content.Shared.Actions;
-
-namespace Content.Shared.Wieldable;
-
-public sealed partial class TwoHandWieldingActionEvent : InstantActionEvent { }
